// src/types.ts
// src/types.ts
export type EBOOKSData = {
    id: number;
    title: string;
    description: string;
    image: string;
    prix: string;
    url: string; // Lien vers le ebook
    // autres propriétés...
  };
  
export type PROJETSData = {
    id: number;
    title: string;
    description: string;
    image: string;
    prix: string;
    url: string; // Lien vers le ebook
    // autres propriétés...
  };
  
export type LOGICIELSData = {
    id: number;
    title: string;
    description: string;
    image: string;
    prix: string;
    url: string; // Lien vers le ebook
    // autres propriétés...
  };
  
export type PSDData = {
    id: number;
    title: string;
    description: string;
    image: string;
    prix: string;
    url: string; // Lien vers le ebook
    // autres propriétés...
  };

  
  export type RCHRETIENNESData = {
    id: number;
    title: string;
    description: string;
    image: string;
    prix: string;
    url: string; // Lien vers le ebook
    // autres propriétés...
  };


  export type VIDEOSData = {
    id: number;
    title: string;
    description: string;
    image: string;
    prix: string;
    url: string; // Lien vers le ebook
    // autres propriétés...
  };
  